requirements.txt
click==8.1.7
Flask==2.2.2
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==23.3.1
python-dotenv==0.21.0
setuptools==68.0.0
Werkzeug==2.2.3
wheel==0.41.2
